Throw error for zero-shard INSERTs
Found this when doing CitusDB integration. If CitusDB has created an
e.g. range partition table I found it "just works" with pg_shard save
one caveat: if no shard exists to receive an incoming row, our zero-
shard SELECT logic kicked in and pg_shard would happily put the row
into the local relation.

UPDATE and DELETE are safe to run locally as the local table should be
empty to begin with, but INSERT is not: we should raise an error rather
than continue with a zero-shard INSERT.