SPI-ify Partition metadata functions
Easiest to start with, with one caveat: the SELECT statements needed to
implement these functions with SPI themselves trigger pg_shard code! So
in IsDistributedTable we need to short-circuit if a target table is in
the pgs_distribution_metadata schema, otherwise we end up with infinite
recursion.

Another fun thing is that SPI runs everything within a special memory
context that it manages. It provides an SPI_palloc function to reach
out to the top-level context, but that only helps replace direct palloc
calls, not calls to system functions that themselves use palloc. It
seems to me that manually switching to another context is the cleanest
way around this "feature" of SPI.